requirements.txt
requests
google-generativeai
supabase